just keep a record before making a big change
Currently, it works for only training the critic. If I train the actor, the error is about 15%-20%, but the objective function function is nearly optimal. Therefore, we can try to train both (ADMM). I am going to change the code s.t. the actor and critic have different parameter for time and #interval. The critic usually need more #time interval while the actor does not. What's more, since the control in the propagation process, the training step for actor is much more expensive.